graphic_chat/requirements.txt
aiofiles==0.4.0
python-dotenv==0.19.0
async-timeout==5.0.1
anyio==4.9.0
exceptiongroup==1.2.2
